core: allow URL parameters and POST bodies to co-exist
Node 8.14.0 prohibits HTTP headers that exceed 8kB.  This patch
allows for the parameters within the body of an HTTP POST request
to be used in addition to those within the URL (and will override
them).

Alternatively for POST we could ignore URL parameters completely,
but the relevant backend tests would need to be rewritten so as
not to use `endPoint('funcName')` which generates a URL that
pre-includes the API key as a request parameter.